agent-python/requirements.txt
openai-agents
python-dotenv
anthropic
httpx
